fix: move samplProducts to a separate file and change imports
