+++ app/(app)/bookmarked-products/[folderId]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { sampleProducts } from "@/app/(app)/products/page";
+import { sampleProducts } from "@/app/(app)/products/data";
 import { useParams, useRouter } from "next/navigation";
 import { Card, CardContent } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
